+++ model/PageDisplay.ts
@@ -1,0 +1,5 @@
+export enum PageDisplay {
+  OnePage = "OnePage",
+  TwoPagesWithCover = "TwoPagesWithCover",
+  TwoPagesNoCover = "TwoPagesNoCover",
+}