+++ model/ReaderPage.ts
@@ -1,0 +1,8 @@
+import {Dimensions} from "./Dimensions"
+import {OcrPage} from "./OcrPage"
+
+export interface ReaderPage {
+  readonly index: number,
+  readonly dimensions: Dimensions,
+  readonly ocr: OcrPage,
+}