+++ pages/read/[bookId]/[page].tsx
@@ -1,16 +1,15 @@
 import PageHead, {defaultPageTitle} from "../../../components/PageHead"
-import {Box, Flex, Grid, GridItem, HStack, Image} from "@chakra-ui/react"
+import {Box, Flex, Grid, GridItem, HStack, IconButton, Image, Tooltip} from "@chakra-ui/react"
 import React, {useEffect, useRef, useState} from "react"
 import {useRouter} from "next/router"
 import {ColorModeSwitcher} from "../../../components/ColorModeSwitcher"
 import {services} from "../../../service/Services"
 import {GetServerSidePropsContext} from "next"
-import {OcrPage} from "../../../model/OcrPage"
 import ZoomSelector from "../../../components/ZoomSelector"
 import SelectionColorOverride from "../../../components/SelectionColorOverride"
 import PageSwitcher from "../../../components/PageSwitcher"
-import {readBookRoute} from "../../../util/Route"
-import {bookAnalyzePageUrl, bookPageUrl} from "../../../util/Url"
+import {notFoundRoute, readBookRoute} from "../../../util/Route"
+import {bookPageUrl} from "../../../util/Url"
 import SvgOverlay from "../../../components/SvgOverlay"
 import {ParsedUrlQuery} from "querystring"
 import ReaderMenu from "../../../components/ReaderMenu"
@@ -19,7 +18,6 @@ import {ReaderSettings} from "../../../model/ReaderSettings"
 import ReadingTimer from "../../../components/ReadingTimer"
 import {ReadingUnitType} from "../../../model/ReadingUnitType"
 import {Api} from "../../../util/Api"
-import {Dimensions} from "../../../model/Dimensions"
 import {ImageAnalysis} from "../../../model/ImageAnalysis"
 import {JpdbRule} from "../../../model/JpdbRule"
 import {PopupPosition} from "../../../model/PopupPosition"
@@ -27,12 +25,19 @@ import {PageViewWrapper, PageViewWrapperHandle} from "../../../components/PageVi
 import {PageView} from "../../../model/PageView"
 import {FloatingPageSettings} from "../../../model/AppSettings"
 import {useImmer} from "use-immer"
+import {MdSettingsBackupRestore} from "react-icons/md"
+import {TOOLTIP_OPEN_DELAY} from "../../../util/Util"
+import {calculatePageStep, calculateWantedPages} from "../../../util/Pages"
+import {ReaderPage} from "../../../model/ReaderPage"
+import {ReadingDirection} from "../../../model/ReadingDirection"
+import {PageDisplay} from "../../../model/PageDisplay"
 
 interface Props {
   title: string,
-  pages: number,
-  ocr: OcrPage,
-  pageDimensions: Dimensions,
+  totalPages: number,
+  pages: readonly ReaderPage[],
+  lowPage: number,
+  highPage: number,
   jpdbRules: readonly JpdbRule[],
   jpdbEnabled: boolean,
   jpdbMiningDeckId: number,
@@ -54,9 +59,10 @@ function getParams(params: ParsedUrlQuery) {
 export default function ReadBookPage(
   {
     title,
+    totalPages,
     pages,
-    ocr,
-    pageDimensions,
+    lowPage,
+    highPage,
     jpdbRules,
     jpdbEnabled,
     jpdbMiningDeckId,
@@ -68,17 +74,18 @@ export default function ReadBookPage(
   }: Props
 ) {
   const router = useRouter()
-  const {bookId, page} = getParams(router.query)
+  const {bookId} = getParams(router.query)
 
   const [readerSettings, updateReaderSettings] = useImmer(initialReaderSettings)
 
   const [fontSizeHover, setFontSizeHover] = useState(false)
   const [minimumConfidenceHover, setMinimumConfidenceHover] = useState(false)
-  const [analysis, setAnalysis] = useState<ImageAnalysis | undefined>(undefined)
+  const [analysis, setAnalysis] = useState<(ImageAnalysis | undefined)[] | undefined>(undefined)
   const [analysisStarted, setAnalysisStarted] = useState(false)
+  const [reloadRequired, setReloadRequired] = useState(false)
 
   const [charactersRead, setCharactersRead] = useState(0)
-  const [charactersReadMaxPage, setCharactersReadMaxPage] = useState(page)
+  const [charactersReadMaxPage, setCharactersReadMaxPage] = useState(lowPage)
   const [pagesRead, setPagesRead] = useState(0)
 
   const pageViewWrapperRef = useRef<PageViewWrapperHandle>(null)
@@ -86,9 +93,13 @@ export default function ReadBookPage(
   useEffect(() => {
     const timer = setTimeout(async () => {
       await Api.updateReaderSettings(bookId, readerSettings)
+      if (reloadRequired) {
+        setReloadRequired(false)
+        await router.push(readBookRoute(bookId, lowPage + 1))
+      }
     }, 300)
     return () => clearTimeout(timer)
-  }, [bookId, readerSettings])
+  }, [bookId, lowPage, readerSettings, reloadRequired, router])
 
   useEffect(() => {
     pageViewWrapperRef.current?.zoomToPageNow()
@@ -99,36 +110,33 @@ export default function ReadBookPage(
       return
     }
     setAnalysisStarted(true)
-    const res = await fetch(bookAnalyzePageUrl(bookId, page - 1))
-    if (res.ok) {
-      setAnalysis(await res.json())
-      updateReaderSettings(draft => {
-        draft.showAnalysis = true
-      })
-    } else {
-      setAnalysisStarted(false)
-      console.log("Failed to analyze page")
-      console.log(res)
-    }
+    setAnalysis(await Api.analyze(bookId, pages.map(it => it.index)))
+    updateReaderSettings(draft => {
+      draft.showAnalysis = true
+    })
   }
 
   async function changePage(newPage: number) {
+    console.log(`Change page to ${newPage}`)
     setAnalysis(undefined)
     setAnalysisStarted(false)
     window.getSelection()?.removeAllRanges()
 
     if (newPage > charactersReadMaxPage) {
-      setCharactersRead(it => it + ocr.characterCount)
+      const characterCount = pages
+        .map(it => it.ocr.characterCount)
+        .reduce((a, b) => a + b, 0)
+      setCharactersRead(it => it + characterCount)
       setCharactersReadMaxPage(newPage)
       setPagesRead(it => it + 1)
-      console.log(`Stats: add ${ocr.characterCount} characters, new max page: ${newPage}`)
+      console.log(`Stats: add ${characterCount} characters, new max page: ${newPage}`)
     }
     await router.push(readBookRoute(bookId, newPage))
   }
 
   function onReadingTimerReset() {
     setCharactersRead(0)
-    setCharactersReadMaxPage(page)
+    setCharactersReadMaxPage(lowPage)
     setPagesRead(0)
   }
 
@@ -144,22 +152,34 @@ export default function ReadBookPage(
             </GridItem>
             <GridItem justifySelf="center">
               <PageSwitcher
-                page={page}
-                pages={pages}
+                lowPage={lowPage + 1}
+                highPage={highPage + 1}
+                totalPages={totalPages}
+                step={calculatePageStep(lowPage, readerSettings.pageDisplay)}
                 readingDirection={readerSettings.readingDirection}
                 onChange={(newPage) => changePage(newPage)}
               />
             </GridItem>
             <GridItem justifySelf="end">
               <HStack>
-                {readerSettings.pageView === PageView.Fixed &&
+                {readerSettings.pageView === PageView.Fixed ?
                   <Box pr={4}>
                     <ZoomSelector
                       zoom={readerSettings.zoom}
                       onChange={v => updateReaderSettings(draft => {
                         draft.zoom = v
                       })}/>
-                  </Box>}
+                  </Box> : <Tooltip label="Reset view" openDelay={TOOLTIP_OPEN_DELAY}>
+                    <IconButton
+                      size="md"
+                      fontSize="lg"
+                      variant="ghost"
+                      color="current"
+                      icon={<MdSettingsBackupRestore/>}
+                      aria-label="Reset view"
+                      onClick={() => pageViewWrapperRef.current?.zoomToPageNow()}
+                    />
+                  </Tooltip>}
                 {readingTimerEnabled ?
                   <ReadingTimer
                     charactersRead={charactersRead}
@@ -182,12 +202,23 @@ export default function ReadBookPage(
                   onChangeTextOrientation={it => updateReaderSettings(draft => {
                     draft.textOrientation = it
                   })}
-                  onChangeReadingDirection={it => updateReaderSettings(draft => {
-                    draft.readingDirection = it
-                  })}
+                  onChangeReadingDirection={it => {
+                    updateReaderSettings(draft => {
+                      draft.readingDirection = it
+                    })
+                    if (readerSettings.pageDisplay !== PageDisplay.OnePage) {
+                      setReloadRequired(true)
+                    }
+                  }}
                   onChangePageView={it => updateReaderSettings(draft => {
                     draft.pageView = it
                   })}
+                  onChangePageDisplay={it => {
+                    updateReaderSettings(draft => {
+                      draft.pageDisplay = it
+                    })
+                    setReloadRequired(true)
+                  }}
                   onAutoFontSizeChange={it => updateReaderSettings(draft => {
                     draft.autoFontSize = it
                   })}
@@ -207,36 +238,43 @@ export default function ReadBookPage(
           </Grid>
           <PageViewWrapper
             pageView={readerSettings.pageView}
-            pageDimensions={pageDimensions}
             zoom={readerSettings.zoom}
             floatingPage={floatingPage}
             ref={pageViewWrapperRef}
             wrapper={(divRef, width, alignSelf) =>
-              <div style={{position: "relative", width: width, alignSelf: alignSelf}} ref={divRef}>
-                <Image
-                  alt="Page"
-                  htmlWidth={pageDimensions.w}
-                  htmlHeight={pageDimensions.h}
-                  src={bookPageUrl(bookId, page - 1)}
-                  onLoad={() => pageViewWrapperRef.current?.pageTurned()}
-                />
-                <SvgOverlay
-                  ocr={ocr}
-                  pageDimensions={pageDimensions}
-                  analysis={analysis}
-                  jpdbRules={jpdbRules}
-                  showParagraphs={readerSettings.showParagraphs || minimumConfidenceHover}
-                  showText={readerSettings.showText || fontSizeHover}
-                  showAnalysis={readerSettings.showAnalysis}
-                  autoFontSize={readerSettings.autoFontSize}
-                  fontSize={readerSettings.fontSize}
-                  textOrientation={readerSettings.textOrientation}
-                  minimumConfidence={readerSettings.minimumConfidence}
-                  jpdbMiningDeckId={jpdbMiningDeckId}
-                  jpdbHorizontalTextPopupPosition={jpdbHorizontalTextPopupPosition}
-                  jpdbVerticalTextPopupPosition={jpdbVerticalTextPopupPosition}
-                />
-              </div>
+              <HStack alignSelf={alignSelf} ref={divRef} spacing={0}>
+                {pages.map((page, index) => {
+                  return <div
+                    key={page.index}
+                    style={{position: "relative", width: width(page.dimensions)}}
+                  >
+                    <Image
+                      alt={`Page ${page.index + 1}`}
+                      htmlWidth={page.dimensions.w}
+                      htmlHeight={page.dimensions.h}
+                      src={bookPageUrl(bookId, page.index)}
+                      onLoad={() => pageViewWrapperRef.current?.pageTurned()}
+                      userSelect="none"
+                    />
+                    <SvgOverlay
+                      ocr={page.ocr}
+                      pageDimensions={page.dimensions}
+                      analysis={analysis && analysis[index]}
+                      jpdbRules={jpdbRules}
+                      showParagraphs={readerSettings.showParagraphs || minimumConfidenceHover}
+                      showText={readerSettings.showText || fontSizeHover}
+                      showAnalysis={readerSettings.showAnalysis}
+                      autoFontSize={readerSettings.autoFontSize}
+                      fontSize={readerSettings.fontSize}
+                      textOrientation={readerSettings.textOrientation}
+                      minimumConfidence={readerSettings.minimumConfidence}
+                      jpdbMiningDeckId={jpdbMiningDeckId}
+                      jpdbHorizontalTextPopupPosition={jpdbHorizontalTextPopupPosition}
+                      jpdbVerticalTextPopupPosition={jpdbVerticalTextPopupPosition}
+                    />
+                  </div>
+                })}
+              </HStack>
             }
           />
         </Flex>
@@ -248,17 +286,45 @@ export default function ReadBookPage(
 export async function getServerSideProps(context: GetServerSidePropsContext) {
   const {bookId, page} = getParams(context.query)
   const pageIndex = page - 1
-  const ocr = await services.bookService.getBookOcrResults(bookId, pageIndex)
-  const pageDimensions = await services.bookService.getBookPageDimensions(bookId, pageIndex)
-  const book = await services.bookService.updateBookProgress(bookId, pageIndex)
+  if (pageIndex < 0) {
+    return {redirect: {destination: notFoundRoute(), permanent: false}}
+  }
+  const book = await services.bookService.getBookById(bookId)
   const readerSettings = await services.settingsService.getReaderSettings(book)
   const appSettings = await services.settingsService.getAppSettings()
+  const totalPages = book.images.length
+  const wantedPages = calculateWantedPages(pageIndex, totalPages, readerSettings.pageDisplay)
+
+  if (wantedPages.length === 0) {
+    return {redirect: {destination: notFoundRoute(), permanent: false}}
+  }
+  if (wantedPages[0] !== pageIndex) {
+    return {redirect: {destination: readBookRoute(bookId, wantedPages[0] + 1), permanent: false}}
+  }
+
+  const lowPage = wantedPages[0]
+  const highPage = wantedPages[wantedPages.length - 1]
+  await services.bookService.updateBookProgress(bookId, highPage)
+  if (readerSettings.readingDirection === ReadingDirection.RightToLeft) {
+    wantedPages.reverse()
+  }
+  const pages = await Promise.all(wantedPages.map(async (pageIndex): Promise<ReaderPage> => {
+    const ocr = await services.bookService.getBookOcrResults(bookId, pageIndex)
+    const dimensions = await services.bookService.getBookPageDimensions(bookId, pageIndex)
+    return {
+      index: pageIndex,
+      dimensions: dimensions,
+      ocr: ocr,
+    }
+  }))
+
   return {
     props: {
       title: book.title,
-      pages: book.images.length,
-      ocr: ocr,
-      pageDimensions: pageDimensions,
+      totalPages: totalPages,
+      lowPage: lowPage,
+      highPage: highPage,
+      pages: pages,
       jpdbEnabled: await services.jpdbService.isEnabled(),
       jpdbRules: appSettings.jpdbRules,
       jpdbMiningDeckId: appSettings.jpdbMiningDeckId,