+++ CHANGES.md
@@ -1,5 +1,6 @@
 #### Version 1.4
 - Added alternative page view mode where page can be freely panned and zoomed.
+- Added alternative two-page display mode.
 - Font size and text orientation options are now hidden by default in the reader menu. Hold Alt key to show them.
 
 #### Version 1.3.3